Rename "fsm" to "session"
