+++ tests/test_fsm_validate.py
@@ -33,8 +33,8 @@ from rflx.expression import (
     Variable,
     VariableDeclaration,
 )
-from rflx.fsm import State, StateMachine, StateName, Transition
 from rflx.identifier import ID
+from rflx.session import State, StateMachine, StateName, Transition
 from rflx.statement import Assignment, Erase, Reset
